m1

+++ pybricks/pyBricks1/main.py
@@ -14,34 +14,36 @@ def main():
     # when getting vegetable, yung bigat could affect the gyro, so make sure to compute distance
     # via sensing the white color or ibangga sa edge
     # get red vegetable and yellow
-    WheelController.move_wheels_forward_in_straight_line(float(105))
+    WheelController.move_wheels_forward_in_straight_line(float(90))
     WheelController.wheel_left_turn()
-    WheelController.move_wheels_forward_in_straight_line(float(500))
+    WheelController.move_wheels_forward_in_straight_line(float(260))
     GripperController.grip_element_using_both_arms()
 
     # get another set of vegetable
-    WheelController.move_wheels_backward_in_straight_line(float(200))
+    WheelController.move_wheels_backward_in_straight_line(float(120))
     WheelController.wheel_right_turn()
-    WheelController.move_wheels_forward_in_straight_line(float(250))
+    WheelController.move_wheels_forward_in_straight_line(float(140))
     WheelController.wheel_left_turn()
     GripperController.release_element_using_both_arms()
-    WheelController.move_wheels_forward_in_straight_line(float(200))
+    WheelController.move_wheels_forward_in_straight_line(float(120))
     GripperController.grip_element_using_both_arms()
 
-    WheelController.move_wheels_backward_in_straight_line(float(500))
+    # going long straight
+    WheelController.move_wheels_backward_in_straight_line(float(160))
     WheelController.wheel_right_turn()
     WheelController.move_wheels_forward_in_straight_line(float(1600))
     WheelController.wheel_slight_left_turn()
-    WheelController.move_wheels_forward_in_straight_line(float(400))
+    WheelController.move_wheels_forward_in_straight_line(float(230))
     GripperController.reset_left_arm()
-    WheelController.move_wheels_backward_in_straight_line(float(400))
+    WheelController.move_wheels_backward_in_straight_line(float(230))
     WheelController.wheel_slight_left_turn()
     GripperController.grip_element_using_left_arm()
 
-    WheelController.move_wheels_backward_in_straight_line(float(1200))
-    WheelController.move_wheels_forward_in_straight_line(float(100))
+    # going to red market
+    WheelController.move_wheels_backward_in_straight_line(float(800))
+    WheelController.move_wheels_forward_in_straight_line(float(120))
     WheelController.wheel_right_turn()
-    WheelController.move_wheels_backward_in_straight_line(float(400))
+    WheelController.move_wheels_backward_in_straight_line(float(600))
     GripperController.release_element_using_both_arms()
 
     # go back to starting point
